requirements.txt
docker-registry-core>=2,<3
boto>=2.32.0,<3